Chore: Users 테이블의 TotalIncome 속성을 제거한다

+++ My/Domain/DTO/ProfileResponse.cs
@@ -9,6 +9,6 @@
         public required string? AccountNumber { get; set; }
         public required string PhoneNumber { get; set; }
         public required float MannerScore { get; set; }
-        public required int TotalIncome { get; set; }
+        public required decimal TotalIncome { get; set; }
     }
 }
